Add files via upload
The Problem

Captain Jack and his crew of pirates arrive in Tortuga after several weeks of plundering the high seas. Since the crew has been at sea for nearly a month, they are ready for a night of celebration. Captain Jack doesn't want to divvy up the treasure just yet, so he gives each man in the crew (except for himself and the first mate) 3 pieces of gold and sends them into town to celebrate. After the crew has gone he and the first mate count what's left and decide how to split it up among the crew. Captain Jack decides that he'll take 12% of the gold. He counts out his gold and hides it in his cabin. The Captain agrees to give the first mate 8% of the remaining gold. The first mate counts out his coins and hides it in his cabin. The next morning, the gold that is left is divided evenly among the members of the crew, including Captain Jack and the First Mate. Little do they know that Cap'n Jack and the first mate have already taken a cut. If the remaining treasure can't be split